lp/requirements.txt
Flask>=2.3.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
requests>=2.28.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=21.0.0,<22.0.0
stripe>=6.0.0,<7.0.0
Pillow>=9.0.0,<10.0.0
schedule>=1.2.0,<2.0.0
psutil>=5.9.0,<6.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
PyJWT>=2.8.0,<3.0.0
gspread>=5.10.0,<6.0.0
google-auth>=2.17.0,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.1.0,<1.0.0
